pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "passport"
version = "2.4.3"
description = "User identification service"
authors = ["Kirill Sumorokov <sumorokov.k@gmail.com>"]
license = "MIT"
packages = [
    { include = "passport", from = "src" },
]

[tool.poetry.dependencies]
aiohttp-metrics = {git = "https://github.com/clayman74/aiohttp-metrics.git", rev = "v0.1.2"}
aiohttp-micro = {git = "https://github.com/clayman74/aiohttp-micro.git", rev = "v0.2.4"}
asyncpg = "0.20.1"
passlib = "1.7.2"
pyjwt = {extras = ["crypto"], version = "1.7.1"}
python = "^3.8"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
coverage = "*"
faker = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-builtins-unleashed = "*"
flake8-comprehensions = "*"
flake8-import-order = "*"
flake8-print = "*"
flake8-pytest = "*"
mypy = "*"
pytest = "*"
pytest-aiohttp = "*"
pytest-postgres = "*"
rope = "*"
ipython = "^7.12.0"
